Edit StyleSheet, Edit IndexHTML, add Icons
This commit will edit the stylesheet so that everything is more readable and the color palette is less harsh.

It adds icons to the side nav bar and changes the size of the side bar blocks.

It adds a headshot of myself to the header on the index page.

It changes the font sizes of the h1 and h2 elements.

The text in the body of the website is edited as well to include a better summary of myself.